+++ src/main/java/de/samply/annotations/EmailSender.java
@@ -3,11 +3,9 @@ package de.samply.annotations;
 import de.samply.notification.smtp.EmailRecipientType;
 import de.samply.notification.smtp.EmailTemplateType;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
+@Repeatable(EmailSenders.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface EmailSender {